refactor(deploy): split orchestration.rs god-file into modules
