+++ src/core/deploy/orchestration/mod.rs
@@ -3,28 +3,28 @@ use std::collections::HashMap;
 use crate::core::component::Component;
 use crate::core::context::RemoteProjectContext;
 use crate::core::error::{Error, Result};
-use crate::core::git;
 use crate::core::project::Project;
 use crate::core::release::version;
 
 use super::execution::{
-    execute_preflighted_component_deploy, prepare_component_deploy, release_artifact_plan,
-    PreparedComponentDeploy, ReleaseArtifactPlan,
+    execute_preflighted_component_deploy, prepare_component_deploy, PreparedComponentDeploy,
 };
-use super::generated_artifacts::uncommitted_file_report_excluding_known_generated;
-use super::orchestration_tag_checkout::{
-    checkout_deploy_tags, deploy_tag_for_version, restore_branches, TagCheckout,
-};
+use super::orchestration_tag_checkout::{checkout_deploy_tags, restore_branches};
 use super::path_roots::{project_with_detected_path_roots, resolve_effective_remote_path};
-use super::planning::{
-    calculate_component_status_with_git_cache, calculate_release_state, load_project_components,
-    plan_components, ExtensionSkippedComponent, GitProbeCache,
-};
-use super::types::{
-    ComponentDeployResult, ComponentStatus, DeployArtifactSource, DeployConfig,
-    DeployOrchestrationResult, DeploySummary,
-};
+use super::planning::{load_project_components, plan_components};
+use super::types::{ComponentDeployResult, DeployConfig, DeployOrchestrationResult, DeploySummary};
 use super::version_overrides::fetch_remote_versions_for_project;
+
+mod modes;
+mod preflight;
+mod smoke_check;
+
+use modes::{extension_skipped_results, run_check_mode, run_dry_run_mode};
+use preflight::{
+    check_uncommitted_changes, check_unreleased_commits, sync_components, verify_expected_version,
+    warn_non_default_branch,
+};
+use smoke_check::run_post_deploy_smoke;
 
 /// Main deploy orchestration entry point.
 /// Handles component selection, building, and deployment.
@@ -280,54 +280,6 @@ pub(super) fn deploy_components(
     })
 }
 
-/// Run the project's post-deploy smoke check, recording the outcome on the
-/// deploy results.
-///
-/// Returns:
-/// - `None` when no smoke check is configured/enabled,
-/// - `Some(true)` when the smoke FAILED and should fail the deploy,
-/// - `Some(false)` when the smoke passed or only warned.
-///
-/// Warnings/errors are appended to the first deployed component result so they
-/// surface in CLI/JSON output alongside the deploy that triggered them.
-fn run_post_deploy_smoke(project: &Project, results: &mut [ComponentDeployResult]) -> Option<bool> {
-    let config = project.smoke_check.as_ref()?;
-    let outcome = super::smoke::run_smoke_check(config)?;
-
-    if outcome.is_ok() {
-        log_status!(
-            "deploy",
-            "Post-deploy smoke check passed for '{}' ({})",
-            project.id,
-            config.url
-        );
-        return Some(false);
-    }
-
-    let detail = outcome
-        .failure_detail()
-        .unwrap_or("post-deploy smoke check failed")
-        .to_string();
-
-    if config.warn_only {
-        log_status!("deploy", "Warning: {} (warn_only)", detail);
-        if let Some(first) = results.iter_mut().find(|r| r.status == "deployed") {
-            first.warnings.push(format!("{} (warn_only)", detail));
-        }
-        return Some(false);
-    }
-
-    log_status!(
-        "deploy",
-        "{} — failing deploy; roll back the release",
-        detail
-    );
-    if let Some(first) = results.iter_mut().find(|r| r.status == "deployed") {
-        first.error = Some(detail);
-    }
-    Some(true)
-}
-
 fn validate_supported_build_configs(components: &[Component]) -> Result<()> {
     for component in components {
         component.validate_supported_build_config()?;
@@ -383,586 +335,27 @@ fn validate_effective_remote_paths(
     Ok(())
 }
 
-/// Check mode: return component status without building or deploying.
-fn run_check_mode(
-    components: &[Component],
-    local_versions: &HashMap<String, String>,
-    remote_versions: &HashMap<String, String>,
-    extension_skipped: &[ExtensionSkippedComponent],
-    project: &Project,
-    base_path: &str,
-    config: &DeployConfig,
-) -> DeployOrchestrationResult {
-    let mut git_probe_cache = GitProbeCache::default();
-    let mut results: Vec<ComponentDeployResult> = components
-        .iter()
-        .map(|c| {
-            let status =
-                calculate_component_status_with_git_cache(c, remote_versions, &mut git_probe_cache);
-            let release_state = calculate_release_state(c);
-            let mut result = ComponentDeployResult::new_for_project(c, project, base_path)
-                .with_status("checked")
-                .with_versions(
-                    local_versions.get(&c.id).cloned(),
-                    remote_versions.get(&c.id).cloned(),
-                )
-                .with_component_status(status)
-                .with_source_identity(c, config.head);
-            if let Some(state) = release_state {
-                result = result.with_release_state(state);
-            }
-            result
-        })
-        .collect();
-
-    // Append components skipped because a required extension is not installed, so the
-    // check-mode diff reports per-component status for the whole project (issue #4587).
-    let skipped_results = extension_skipped_results(extension_skipped, project, base_path);
-    let skipped = skipped_results.len() as u32;
-    results.extend(skipped_results);
-
-    let total = results.len() as u32;
-    DeployOrchestrationResult {
-        results,
-        summary: DeploySummary {
-            total,
-            succeeded: 0,
-            failed: 0,
-            skipped,
-        },
-    }
-}
-
-/// Build check-mode result rows for components skipped due to missing extensions.
-///
-/// Each row is `status: "skipped"` with a warning explaining the missing extension,
-/// so operators see `skipped: missing extension <id>` instead of the whole pass aborting.
-fn extension_skipped_results(
-    extension_skipped: &[ExtensionSkippedComponent],
-    project: &Project,
-    base_path: &str,
-) -> Vec<ComponentDeployResult> {
-    extension_skipped
-        .iter()
-        .map(|skip| {
-            let component = Component {
-                id: skip.id.clone(),
-                ..Default::default()
-            };
-            let mut result = ComponentDeployResult::new_for_project(&component, project, base_path)
-                .with_status("skipped");
-            result.warnings.push(format!("skipped: {}", skip.reason));
-            result
-        })
-        .collect()
-}
-
-/// Dry-run mode: return planned results without building or deploying.
-fn run_dry_run_mode(
-    components: &[Component],
-    local_versions: &HashMap<String, String>,
-    remote_versions: &HashMap<String, String>,
-    project: &Project,
-    base_path: &str,
-    config: &DeployConfig,
-) -> Result<DeployOrchestrationResult> {
-    let mut git_probe_cache = GitProbeCache::default();
-    let results: Vec<ComponentDeployResult> = components
-        .iter()
-        .map(|c| {
-            let status = if config.check {
-                calculate_component_status_with_git_cache(c, remote_versions, &mut git_probe_cache)
-            } else {
-                ComponentStatus::Unknown
-            };
-            let mut result = ComponentDeployResult::new_for_project(c, project, base_path)
-                .with_status("planned")
-                .with_versions(
-                    local_versions.get(&c.id).cloned(),
-                    remote_versions.get(&c.id).cloned(),
-                )
-                .with_source_identity(c, config.head);
-            if let Some(deploy_ref) = planned_deploy_ref(c, config)? {
-                result = result.with_deployed_ref(deploy_ref);
-            }
-            result = with_dry_run_artifact_plan(result, c, config);
-            if config.check {
-                result = result.with_component_status(status);
-            }
-            Ok(result)
-        })
-        .collect::<Result<Vec<_>>>()?;
-
-    let total = results.len() as u32;
-    Ok(DeployOrchestrationResult {
-        results,
-        summary: DeploySummary {
-            total,
-            succeeded: 0,
-            failed: 0,
-            skipped: 0,
-        },
-    })
-}
-
-fn with_dry_run_artifact_plan(
-    mut result: ComponentDeployResult,
-    component: &Component,
-    config: &DeployConfig,
-) -> ComponentDeployResult {
-    let deploy_config = component.deploy_config();
-    let is_git_deploy = deploy_config.is_git_deploy();
-    let is_file_deploy = deploy_config.is_file_deploy();
-    if is_git_deploy || is_file_deploy {
-        return result;
-    }
-
-    match release_artifact_plan(component, config, is_git_deploy, is_file_deploy) {
-        ReleaseArtifactPlan::Reuse { url, tag } => {
-            result.warnings.push(format!(
-                "artifact source: release asset for tag {tag}; build phase: skipped if asset is available; deploy phase: would upload downloaded asset"
-            ));
-            result
-                .with_artifact_path(Some(url))
-                .with_artifact_source(DeployArtifactSource::ReleaseAsset)
-        }
-        ReleaseArtifactPlan::LocalBuild { reason } => {
-            result.warnings.push(format!(
-                "artifact source: local rebuild; reason: {reason}; build phase: would run before deploy; deploy phase: would upload local build_artifact"
-            ));
-            result.with_artifact_source(DeployArtifactSource::LocalBuild)
-        }
-    }
-}
-
-fn planned_deploy_ref(component: &Component, config: &DeployConfig) -> Result<Option<String>> {
-    if component.is_file_component() {
-        return Ok(None);
-    }
-
-    let path = &component.local_path;
-    if config.head {
-        return Ok(crate::core::engine::command::run_in_optional(
-            path,
-            "git",
-            &["rev-parse", "--abbrev-ref", "HEAD"],
-        )
-        .map(|branch| format!("{} (HEAD)", branch)));
-    }
-
-    let tag = latest_deploy_tag(component, config.expected_version.as_deref())?;
-    let tag_sha =
-        crate::core::engine::command::run_in_optional(path, "git", &["rev-parse", "--short", &tag]);
-    let head_ahead = crate::core::engine::command::run_in_optional(
-        path,
-        "git",
-        &["rev-list", "--count", &format!("{}..HEAD", tag)],
-    )
-    .and_then(|out| out.trim().parse::<u32>().ok())
-    .unwrap_or(0);
-
-    Ok(Some(
-        TagCheckout {
-            component_id: component.id.clone(),
-            tag,
-            original_ref: String::new(),
-            local_path: path.clone(),
-            tag_sha,
-            head_ahead,
-        }
-        .provenance_ref(),
-    ))
-}
-
-fn latest_deploy_tag(component: &Component, expected_version: Option<&str>) -> Result<String> {
-    if let Some(version) = expected_version {
-        return Ok(deploy_tag_for_version(component, version));
-    }
-
-    match git::get_latest_tag(&component.local_path) {
-        Ok(Some(tag)) => Ok(tag),
-        Ok(None) => Err(Error::validation_invalid_argument(
-            "deploy",
-            format!(
-                "Refusing to deploy '{}': no version tags found for default tagged deploy",
-                component.id
-            ),
-            None,
-            Some(vec![
-                "Run `homeboy release` to create a tagged release first".to_string(),
-                "Use `homeboy deploy --head` to deploy the current branch HEAD explicitly"
-                    .to_string(),
-            ]),
-        )),
-        Err(err) => Err(Error::git_command_failed(format!(
-            "Could not read version tags for '{}': {}",
-            component.id, err
-        ))),
-    }
-}
-
-/// Verify no components have uncommitted changes before deployment.
-/// Warn when `--head` would deploy from a non-default branch.
-///
-/// Detects the current branch for each component and compares it against the
-/// default branch (via `git symbolic-ref refs/remotes/origin/HEAD`, falling
-/// back to "main"). If a component is on a feature branch, this is likely
-/// unintentional — the user probably meant to deploy the default branch.
-///
-/// With `--force`, this emits a log warning but proceeds. Without `--force`,
-/// it returns an error so the user can switch branches or confirm intent.
-fn warn_non_default_branch(components: &[Component], config: &DeployConfig) -> Result<()> {
-    for component in components {
-        if component.is_file_component() {
-            continue;
-        }
-
-        let path = &component.local_path;
-
-        // Get current branch
-        let current_branch = match crate::core::engine::command::run_in_optional(
-            path,
-            "git",
-            &["rev-parse", "--abbrev-ref", "HEAD"],
-        ) {
-            Some(branch) if branch != "HEAD" => branch, // "HEAD" means detached
-            _ => continue,                              // detached or error — skip
-        };
-
-        // Detect default branch from remote HEAD symref, fallback to "main"
-        let default_branch = crate::core::engine::command::run_in_optional(
-            path,
-            "git",
-            &["symbolic-ref", "--short", "refs/remotes/origin/HEAD"],
-        )
-        .map(|s| {
-            // Output is like "origin/main" — strip the remote prefix
-            s.strip_prefix("origin/").unwrap_or(&s).to_string()
-        })
-        .unwrap_or_else(|| "main".to_string());
-
-        if current_branch != default_branch {
-            let message = format!(
-                "Component '{}' is on branch '{}', not '{}' (default)",
-                component.id, current_branch, default_branch
-            );
-
-            if config.force {
-                log_status!("deploy", "Warning: {}", message);
-            } else {
-                return Err(Error::validation_invalid_argument(
-                    "head",
-                    message,
-                    None,
-                    Some(vec![
-                        format!(
-                            "Switch to the default branch: git -C {} checkout {}",
-                            component.local_path, default_branch
-                        ),
-                        "Use --force to deploy from the current branch anyway".to_string(),
-                    ]),
-                ));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn check_uncommitted_changes(components: &[Component]) -> Result<()> {
-    // Partition components into "non-git local_path" vs "dirty git repo" so we can
-    // emit the right diagnostic. Conflating the two leaves users chasing a
-    // nonexistent uncommitted-changes problem when the real issue is that
-    // local_path doesn't point at a git checkout. (#1141)
-    let mut non_git: Vec<&Component> = Vec::new();
-    let mut dirty: Vec<DirtyComponent> = Vec::new();
-
-    for component in components {
-        if component.is_file_component() {
-            continue;
-        }
-        if !git::is_git_repo(&component.local_path) {
-            non_git.push(component);
-            continue;
-        }
-        match uncommitted_file_report_excluding_known_generated(component) {
-            Ok(report) if report.unexpected.is_empty() => {}
-            Ok(report) => dirty.push(DirtyComponent {
-                id: component.id.clone(),
-                unexpected_paths: report.unexpected,
-                known_generated_paths: report.known_generated,
-            }),
-            Err(_) => dirty.push(DirtyComponent {
-                id: component.id.clone(),
-                unexpected_paths: Vec::new(),
-                known_generated_paths: Vec::new(),
-            }),
-        }
-    }
-
-    if !non_git.is_empty() {
-        let ids: Vec<&str> = non_git.iter().map(|c| c.id.as_str()).collect();
-        let mut hints: Vec<String> = non_git
-            .iter()
-            .map(|c| {
-                format!(
-                    "Repoint '{}' at a git checkout: homeboy component set {} --local-path <path-to-git-workspace>",
-                    c.id, c.id
-                )
-            })
-            .collect();
-        hints.push(
-            "Initialize a git repo at the existing local_path if the contents are the source of truth: git -C <local_path> init && git add . && git commit -m 'initial'"
-                .to_string(),
-        );
-        hints.push("Or deploy with --force to bypass the git-clean check".to_string());
-        let paths: Vec<String> = non_git
-            .iter()
-            .map(|c| format!("{} ({})", c.id, c.local_path))
-            .collect();
-        return Err(Error::validation_invalid_argument(
-            "components",
-            format!(
-                "local_path is not a git repository for: {}. The uncommitted-changes check requires a git checkout.",
-                paths.join(", ")
-            ),
-            Some(ids.join(",")),
-            Some(hints),
-        ));
-    }
-
-    if !dirty.is_empty() {
-        let dirty_ids: Vec<&str> = dirty.iter().map(|row| row.id.as_str()).collect();
-        return Err(Error::validation_invalid_argument(
-            "components",
-            format!(
-                "Components have uncommitted changes: {}. {}",
-                dirty_ids.join(", "),
-                dirty_worktree_path_summary(&dirty)
-            ),
-            None,
-            Some(vec![
-                "Commit your changes before deploying to ensure deployed code is tracked"
-                    .to_string(),
-                "Known generated artifacts are ignored by the deploy dirty gate; remove them with `homeboy cleanup --apply` if you want a clean worktree"
-                    .to_string(),
-                "Use --force to deploy anyway".to_string(),
-            ]),
-        ));
-    }
-    Ok(())
-}
-
-struct DirtyComponent {
-    id: String,
-    unexpected_paths: Vec<String>,
-    known_generated_paths: Vec<String>,
-}
-
-fn dirty_worktree_path_summary(dirty: &[DirtyComponent]) -> String {
-    dirty
-        .iter()
-        .map(|row| {
-            let unexpected = if row.unexpected_paths.is_empty() {
-                "<unable to read git status>".to_string()
-            } else {
-                row.unexpected_paths.join(", ")
-            };
-            let known_generated = if row.known_generated_paths.is_empty() {
-                String::new()
-            } else {
-                format!(
-                    "; known generated artifacts ignored: {}",
-                    row.known_generated_paths.join(", ")
-                )
-            };
-            format!(
-                "{} unexpected paths: {}{}",
-                row.id, unexpected, known_generated
-            )
-        })
-        .collect::<Vec<_>>()
-        .join("; ")
-}
-
-/// Fetch and pull latest changes for each component before deploying.
-///
-/// Prevents deploying stale code when the local clone is behind remote.
-/// Runs `git fetch` + `git pull` for each component that has an upstream.
-/// Aborts if pull fails (e.g., merge conflicts).
-fn sync_components(components: &[Component]) -> Result<()> {
-    for component in components {
-        // File components are not git repos — skip sync
-        if component.is_file_component() {
-            continue;
-        }
-
-        let path = &component.local_path;
-
-        let version_before_pull = version::get_component_version(component);
-
-        // Check if behind remote
-        match git::fetch_and_get_behind_count(path) {
-            Ok(Some(behind)) => {
-                log_status!(
-                    "deploy",
-                    "'{}' is {} commit(s) behind remote — pulling...",
-                    component.id,
-                    behind
-                );
-                let pull_result = git::pull(Some(&component.id))?;
-                if !pull_result.success {
-                    return Err(Error::git_command_failed(format!(
-                        "Failed to pull '{}': {}",
-                        component.id,
-                        pull_result.stderr.lines().next().unwrap_or("unknown error")
-                    )));
-                }
-                if let Some(message) = auto_pull_version_drift_message(
-                    component,
-                    version_before_pull.as_deref(),
-                    version::get_component_version(component).as_deref(),
-                ) {
-                    log_status!("deploy", "{}", message);
-                }
-                log_status!("deploy", "'{}' is now up to date", component.id);
-            }
-            Ok(None) => {
-                // Not behind or no upstream — nothing to do
-            }
-            Err(_) => {
-                // git fetch failed — warn but don't block (might be offline)
-                log_status!(
-                    "deploy",
-                    "Warning: could not check remote status for '{}' — deploying local state",
-                    component.id
-                );
-            }
-        }
-    }
-    Ok(())
-}
-
-fn auto_pull_version_drift_message(
-    component: &Component,
-    before: Option<&str>,
-    after: Option<&str>,
-) -> Option<String> {
-    if before == after {
-        return None;
-    }
-
-    Some(format!(
-        "Warning: auto-pull changed '{}' deploy version from {} to {}",
-        component.id,
-        before.unwrap_or("unknown"),
-        after.unwrap_or("unknown")
-    ))
-}
-
-/// Check for unreleased commits ahead of the latest tag.
-///
-/// Checks each component for commits between the latest tag and HEAD.
-/// When found and `--force` is not set, returns an error to prevent
-/// silently deploying stale code. Use `deploy --head` to deploy
-/// unreleased commits, or `homeboy release` to tag them first.
-fn check_unreleased_commits(
-    components: &[Component],
-    config: &DeployConfig,
-) -> crate::core::Result<()> {
-    let mut gaps = Vec::new();
-
-    for component in components {
-        if let Some(gap) = super::provenance::detect_tag_gap(component) {
-            super::provenance::warn_tag_gap(&component.id, &gap, "deploy");
-            gaps.push((component.id.clone(), gap));
-        }
-    }
-
-    if gaps.is_empty() {
-        return Ok(());
-    }
-
-    if config.tagged {
-        log_status!(
-            "deploy",
-            "Deploying from tagged releases (--tagged). Use `deploy --head` to include unreleased commits, or `homeboy release` to tag them."
-        );
-        return Ok(());
-    }
-
-    if config.force {
-        log_status!(
-            "deploy",
-            "Deploying from tagged releases (--force). Use `deploy --head` to include unreleased commits, or `homeboy release` to tag them."
-        );
-        return Ok(());
-    }
-
-    let component_list: Vec<String> = gaps
-        .iter()
-        .map(|(id, gap)| format!("{} ({} commits ahead of {})", id, gap.ahead, gap.tag))
-        .collect();
-
-    Err(crate::core::Error::validation_invalid_argument(
-        "deploy",
-        format!(
-            "Refusing to deploy: HEAD has unreleased commits for: {}",
-            component_list.join(", ")
-        ),
-        None,
-        Some(vec![
-            "Run `homeboy release` to tag the commits first".to_string(),
-            "Use `deploy --head` to deploy unreleased commits directly".to_string(),
-            "Use `deploy --force` to deploy the stale tag anyway".to_string(),
-        ]),
-    ))
-}
-
-/// Verify that component versions match the expected version.
-///
-/// When `--version` is used, ensures the local version of each component
-/// matches the asserted version. This catches cases where the local copy
-/// has a different version than what was just released.
-fn verify_expected_version(components: &[Component], expected: &str) -> Result<()> {
-    let mut mismatches = Vec::new();
-
-    for component in components {
-        match version::get_component_version(component) {
-            Some(local_version) if local_version == expected => {}
-            Some(local_version) => mismatches.push(format!(
-                "'{}': local version is {} (expected {})",
-                component.id, local_version, expected
-            )),
-            None => mismatches.push(format!(
-                "'{}': local version could not be read (expected {})",
-                component.id, expected
-            )),
-        }
-    }
-
-    if !mismatches.is_empty() {
-        return Err(Error::validation_invalid_argument(
-            "version",
-            format!("Version mismatch: {}", mismatches.join("; ")),
-            None,
-            Some(vec![
-                "Pull latest changes: git pull".to_string(),
-                "Or remove --version to deploy the current local version".to_string(),
-            ]),
-        ));
-    }
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::core::component::ComponentScriptsConfig;
+    use crate::core::deploy::orchestration_tag_checkout::{
+        checkout_deploy_tags, deploy_tag_for_version, restore_branches, TagCheckout,
+    };
+    use crate::core::deploy::planning::{load_project_components, ExtensionSkippedComponent};
+    use crate::core::deploy::types::ComponentDeployResult;
     use crate::core::project::ProjectComponentAttachment;
     use crate::test_support::with_isolated_home;
     use std::collections::HashMap;
     use std::path::Path;
     use tempfile::TempDir;
+
+    use super::modes::{extension_skipped_results, run_dry_run_mode};
+    use super::preflight::{
+        auto_pull_version_drift_message, check_uncommitted_changes, check_unreleased_commits,
+        verify_expected_version,
+    };
+    use super::smoke_check::run_post_deploy_smoke;
 
     fn make_component(id: &str, local_path: &str) -> Component {
         Component::new(id.to_string(), local_path.to_string(), String::new(), None)
